:recycle: refactor: ProfileAvatar로 이름 수정
